refactor: extract component to EmailTemplateForm.vue
